Moved logic to utils,graphing & spreadsheet js
