+++ src/scripts/app.js
@@ -1,5 +1,6 @@
-import bb, {line} from 'billboard.js';
-import {duration, inflation, calulateNextYearsPayment,calculateYearsCreditPayments, calculate_monthly_payment,calculate_year_investment, calculate_yearly_div, calculate_monthly_div} from './calculate';
+import * as calc from './calculate';
+import * as graphing from './graphing';
+import {makeHumanReadable} from './utils';
 import {resetTable,generateRow,generateTable} from './spreadsheet';
 import {download_csv,export_table_to_csv} from './csv';
 import '../styles/style.css';
@@ -34,89 +35,17 @@ function toggleTheme(e){
   }
 }
 
-/*
-  Formats data number to comma (,) seperated thousand readable
-*/
-function makeHumanReadable(amount) {
-  return parseFloat(amount).toLocaleString("en-NZ", {
-    maximumFractionDigits: 0,
-  });
-}
 
-function generateChart() {
-  y2axisMaxRange = parseInt(dividendArray[dividendArray.length - 1]);
-  y1axisMaxRange = parseInt(principleArray[principleArray.length - 1]) * 1.1;
-  console.log(
-    "----> Div Max Axis " +
-      y2axisMaxRange +
-      "\n ----> Principle Max Axis :" +
-      y1axisMaxRange
-  );
-  chart = bb.generate({
-    bindto: "#chart",
-    data: {
-      type: line(),
-      columns: [principleArray, dividendArray],
-      axes: {
-        Principle: "y",
-        Principle2: "y2",
-        Dividend: "y3",
-      },
-    },
-    grid: {
-      x: {
-        show: true,
-      },
-      y: {
-        show: true,
-      },
-    },
-    axis: {
-      x: {
-        label: "Year",
-      },
-      y: {
-        label: "Dollers ($)",
-        // max: eval(y1axisMaxRange),
-        // min: 0,
-      } /*
-            y2: {
-              show: true,
-              label: "Principle $",
-            },
-            y3: {
-              show: true,
-              label: "Dividend $",
-              
-              max: y2axisMaxRange*2,
-              min: 0,
-            }*/,
-    },
-    tooltip: {
-      init: {
-        show: false,
-        
-      },
-      doNotHide: false,
-      contents: {
-
-        template: "<table><tbody><tr><th colspan='2'>Year {=TITLE}</th></tr>{{<tr><td class=name><span style='background-color:{=COLOR}; width:10px; height:10px; margin:auto; display:inline-block; align-self:center; margin:0 .5em 0 .5em; border-radius:1px;'></span><span>{=NAME}</span></td><td class=value>{=VALUE}<td></td></tr>}}</tbody></table>"
-      },
-      format: {
-        value: function(value,ratio,id) {
-            return "$"+makeHumanReadable(value);
-        }
-      }
-    },
-  });
-}
 
 function resetGraphArray() {
   principleArray = ["Principle"];
   dividendArray = ["Dividend"];
+  chart.destroy();
 }
 
-
+/*
+/*  Setting up 
+*/
 
 function init() {
   currentTheme = localStorage.getItem("theme");
@@ -148,6 +77,9 @@ function init() {
 
   var hideBtn = document.getElementById("hideMain");
 
+  /*
+    Hide or Show Investment Properties Panel
+  */
   hideBtn.addEventListener("click", function () {
     if (!isFormVisible) {
       document.getElementById("form-wrapper").style.display = "flex";
@@ -160,6 +92,9 @@ function init() {
     }
   });
 
+  /*
+  /*  Calculate Ivestment, Display Graphed results and show Table.
+  */
   var calcBtn = document.getElementById("calculate");
   calcBtn.addEventListener("click", function () {
     if (document.getElementById("tablewrapper").style.visibility == "visible") {
@@ -181,21 +116,12 @@ function init() {
       "monthlyPaymentMax: ",
       +document.getElementById("monthlyPaymentMax")
     );
-    let monthly_payment_max_cap = document.getElementById("monthlyPaymentMax")
-      .value;
-      let inflation_rate = document.getElementById("inflationRate").value;
-      let dividend_yield = document.getElementById("dividendRate").value;
-      let reinvest_dividend = document.getElementById("divReinvest").checked;
-      let period = duration(start_age, end_age);
+    let monthly_payment_max_cap = document.getElementById("monthlyPaymentMax").value;
+    let inflation_rate = document.getElementById("inflationRate").value;
+    let dividend_yield = document.getElementById("dividendRate").value;
+    let reinvest_dividend = document.getElementById("divReinvest").checked;
+    let period = calc.duration(start_age, end_age);
     console.log(start_age.value + "," + end_age.value);
-    let first_years_total_monthly_payments = calculateYearsCreditPayments(
-      monthly_payment,
-      monthly_payment_max_cap,
-      0.0
-    );
-    let second_years_total_monthly_payments =
-      eval(first_years_total_monthly_payments) *
-      eval(1 + monthly_payment_increase / 100.0);
     let total_Credit_Payments = 0.0;
     let current_mo_payment = monthly_payment;
     let current_investment = start_amount;
@@ -205,7 +131,6 @@ function init() {
     var currentYearsPrinciple = 0;
     console.log("=============================== New run ===================");
 
-    var rowsArray = [];
     for (let year = 0; year < period; year++) {
       currentYearsPrinciple = current_investment;
       if (year == 0) {
@@ -218,7 +143,7 @@ function init() {
             monthly_payment_increase +
             " }"
         );
-        current_mo_payment = calculate_monthly_payment(
+        current_mo_payment = calc.calculate_monthly_payment(
           current_mo_payment,
           monthly_payment_max_cap,
           undefined
@@ -235,7 +160,7 @@ function init() {
             " }"
         );
 
-        current_mo_payment = calculate_monthly_payment(
+        current_mo_payment = calc.calculate_monthly_payment(
           current_mo_payment,
           monthly_payment_max_cap,
           monthly_payment_increase
@@ -243,7 +168,7 @@ function init() {
 
         console.debug("After mo payment calc: " + current_mo_payment);
       }
-      var yearsPayments = calculateYearsCreditPayments(
+      var yearsPayments = calc.calculateYearsCreditPayments(
         current_mo_payment,
         monthly_payment_max_cap,
         undefined
@@ -258,7 +183,7 @@ function init() {
           " year in payments credit: " +
           yearsPayments
       );
-      current_investment = calculate_year_investment(
+      current_investment = calc.calculate_year_investment(
         current_investment,
         yearsPayments,
         investment_yield
@@ -266,88 +191,69 @@ function init() {
       investmentAfterYield = current_investment;
       investmentAfterDiv = investmentAfterYield;
       divReturned = eval(
-        calculate_yearly_div(current_investment, dividend_yield)
+        calc.calculate_yearly_div(current_investment, dividend_yield)
       );
       if (reinvest_dividend) {
         current_investment = eval(current_investment) + divReturned;
         investmentAfterDiv = current_investment;
       }
-      /*
-            Below is the table row structure
-            <tr>
-              <th scope="row" >1</th>
-              <td>$10000</td><td>$200</td><td>$2400</td><td>$12400</td><td>$16800</td><td>$160</td><td>$16960</td>
-            </tr>
-          */
-      let decemalPlaces = 0;
-      var newTableRow = document.createElement("tr");
-      var newRowHeader = document.createElement("th");
-      newRowHeader.scope = "row";
-      newRowHeader.innerText = eval(year) + 1;
-      newTableRow.appendChild(newRowHeader);
+      // var newTableRow = document.createElement("tr");
+      // var newRowHeader = document.createElement("th");
+      // newRowHeader.scope = "row";
+      // newRowHeader.innerText = eval(year) + 1;
+      // newTableRow.appendChild(newRowHeader);
 
-      var newRowData0_Principle = document.createElement("td");
-      newRowData0_Principle.innerText = makeHumanReadable(currentYearsPrinciple);
-      var newRowData1_MoCredit = document.createElement("td");
-      newRowData1_MoCredit.innerText = makeHumanReadable(current_mo_payment);
-      var newRowData2_TotalCredit = document.createElement("td");
-      newRowData2_TotalCredit.innerText = makeHumanReadable(yearsPayments);
-      var newRowData3_BeforeYield = document.createElement("td");
-      let sumYearsPrinciple =
-        parseFloat(currentYearsPrinciple) + parseFloat(yearsPayments);
-      newRowData3_BeforeYield.innerText = makeHumanReadable(sumYearsPrinciple);
-      var newRowData4_Afteryield = document.createElement("td");
-      newRowData4_Afteryield.innerText = makeHumanReadable(investmentAfterYield);
-      var newRowData5_DivReturned = document.createElement("td");
-      newRowData5_DivReturned.innerText = makeHumanReadable(divReturned);
-      var newRowData6_InvestDiv = document.createElement("td");
-      newRowData6_InvestDiv.innerText = makeHumanReadable(investmentAfterDiv);
+      let newInvestmentPeriod = {
+        currentYearsPrinciple: currentYearsPrinciple,
+        current_mo_payment: current_mo_payment,
+        yearsPayments: yearsPayments,
+        sumYearsPrinciple: parseFloat(currentYearsPrinciple) + parseFloat(yearsPayments),
+        investmentAfterYield: investmentAfterYield,
+        divReturned: divReturned,
+        investmentAfterDiv: investmentAfterDiv,
+        year: year
+      };
 
-      newTableRow.appendChild(newRowData0_Principle);
-      newTableRow.appendChild(newRowData1_MoCredit);
-      newTableRow.appendChild(newRowData2_TotalCredit);
-      newTableRow.appendChild(newRowData3_BeforeYield);
-      newTableRow.appendChild(newRowData4_Afteryield);
-      newTableRow.appendChild(newRowData5_DivReturned);
-      newTableRow.appendChild(newRowData6_InvestDiv);
+      generateRow(newInvestmentPeriod);
+      // var newRowData0_Principle = document.createElement("td");
+      // newRowData0_Principle.innerText = makeHumanReadable(currentYearsPrinciple);
+      // var newRowData1_MoCredit = document.createElement("td");
+      // newRowData1_MoCredit.innerText = makeHumanReadable(current_mo_payment);
+      // var newRowData2_TotalCredit = document.createElement("td");
+      // newRowData2_TotalCredit.innerText = makeHumanReadable(yearsPayments);
+      // var newRowData3_BeforeYield = document.createElement("td");
+      // let sumYearsPrinciple =
+      //   parseFloat(currentYearsPrinciple) + parseFloat(yearsPayments);
+      // newRowData3_BeforeYield.innerText = makeHumanReadable(sumYearsPrinciple);
+      // var newRowData4_Afteryield = document.createElement("td");
+      // newRowData4_Afteryield.innerText = makeHumanReadable(investmentAfterYield);
+      // var newRowData5_DivReturned = document.createElement("td");
+      // newRowData5_DivReturned.innerText = makeHumanReadable(divReturned);
+      // var newRowData6_InvestDiv = document.createElement("td");
+      // newRowData6_InvestDiv.innerText = makeHumanReadable(investmentAfterDiv);
 
-      document.getElementById("dynamic").appendChild(newTableRow);
+      // newTableRow.appendChild(newRowData0_Principle);
+      // newTableRow.appendChild(newRowData1_MoCredit);
+      // newTableRow.appendChild(newRowData2_TotalCredit);
+      // newTableRow.appendChild(newRowData3_BeforeYield);
+      // newTableRow.appendChild(newRowData4_Afteryield);
+      // newTableRow.appendChild(newRowData5_DivReturned);
+      // newTableRow.appendChild(newRowData6_InvestDiv);
+
+      // document.getElementById("dynamic").appendChild(newTableRow);
+      document.getElementById("dynamic").appendChild(generateRow(newInvestmentPeriod));
+      
       console.debug(
         "[ " + year + " ] After yield added: " + current_investment
       );
       principleArray.push(parseInt(currentYearsPrinciple));
       dividendArray.push(parseInt(divReturned));
-      //console.log("Year "+year+": $"+current_mo_payment);
     }
-    let message =
-      "Investment period of: " +
-      period +
-      " Years" +
-      " Initial Investment: $" +
-      start_amount +
-      " Monthly Payments of: $" +
-      monthly_payment +
-      " Payment Increase rate: " +
-      monthly_payment_increase +
-      "% PA" +
-      "Inflation rate of: " +
-      inflation_rate +
-      "%\n" +
-      "First Years Total Payments: $" +
-      first_years_total_monthly_payments +
-      "\nSecond Years Total Payments: $" +
-      second_years_total_monthly_payments +
-      "\nTotal Credit Payments: $" +
-      total_Credit_Payments +
-      "\nYour Investment is worth: $" +
-      current_investment;
-    //alert(message);
-
     if (chart === null) {
-      generateChart();
+      chart = graphing.generateChart(dividendArray, principleArray);
     } else {
       chart.destroy();
-      generateChart();
+      chart = graphing.generateChart(dividendArray, principleArray);
     }
   });
   
@@ -356,14 +262,13 @@ function init() {
     document.getElementById("tablewrapper").style.visibility = "hidden";
     resetGraphArray();
     resetTable();
+    document.getElementById("divReinvest").checked = false;
     document.getElementById("startAge").value = 21;
     document.getElementById("endAge").value = 65;
     document.getElementById("startAmount").value = 5000;
     document.getElementById("startPaymentAmount").value = 200;
     document.getElementById("monthlyPaymentIncreasePa").value = 0.0;
     document.getElementById("inflationRate").value = 3.0;
-    // console.log(eval(start_age.value) + "," + eval(end_age.value));
-    chart.destroy();
   });
   var csvBtn = document.getElementById("csvfile");
   csvBtn.addEventListener("click", function () {
@@ -380,9 +285,7 @@ function init() {
 }
 
 if (process.env.NODE_ENV !== 'production') {
-
   console.log('Looks like we are in development mode!');
-
 }
 
 window.onload = init();
