+++ src/scripts/utils.js
@@ -1,0 +1,8 @@
+/*
+  Formats data number to comma (,) seperated thousand readable
+*/
+export function makeHumanReadable(amount) {
+    return parseFloat(amount).toLocaleString("en-NZ", {
+      maximumFractionDigits: 0,
+    });
+}
